Слепые псы и общее поведение мутантов из ЗП
Все взято из
https://github.com/xrOxygen/xray-oxygen

+++ ogsr_engine/xrGame/ai/monsters/monster_corpse_memory.h
@@ -18,14 +18,17 @@ public:
 
 	// -----------------------------------------------------
 	const CEntityAlive	*get_corpse				();
+
 	SMonsterCorpse		get_corpse_info			();
-	u32					get_corpse_count		() {return m_objects.size();}
+	u32					get_corpse_count		() {return u32(m_objects.size());}
 
 	void				clear					() {m_objects.clear();}
 	void				remove_links			(CObject *O);
 
 	void				add_corpse				(const CEntityAlive *corpse);
 	bool                is_valid_corpse			(const CEntityAlive *corpse);
+	
+	
 
 private:
 	void				remove_non_actual		();
